pyproject.toml
[project]
name = "perplexity-clone"
version = "0.1.0"
description = "Multi-agent search system with LLM judge for intelligent answers with citations"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",
]

[project.scripts]
perplexity = "src.main:main"
perp = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
